Add order to removeItem

+++ resolvers/removeItem.js
@@ -1,3 +1,5 @@
+const gql = String.raw;
+
 const {getUserId} = require('../utils');
 const {NotFoundError} = require('../errors');
 
@@ -32,13 +34,48 @@ const removeItem = async (parent, {id}, ctx) => {
 				},
 			],
 		},
-	});
+	}).$fragment(gql`
+		fragment ItemWithSectionItems on Item {
+			id
+			status
+			section {
+				project {
+					status
+				}
+				items(orderBy: position_ASC) {
+					id
+					position
+				}
+			}
+		}
+	`);
 
 	if (!item) {
 		throw new NotFoundError(`Item '${id}' has not been found.`);
 	}
 
-	return ctx.db.deleteItem({id});
+	if (
+		item.section.project.status === 'FINISHED'
+		|| item.status === 'FINISHED'
+	) {
+		throw new Error(`Item '${id}' can't be removed in this state.`);
+	}
+
+	const itemIndex = item.section.items.findIndex(
+		sectionItem => item.id === sectionItem.id,
+	);
+
+	const removedItem = await ctx.db.deleteItem({id});
+
+	// updating all the positions from the item position
+	await Promise.all(
+		item.section.items.slice(itemIndex + 1).map((sectionItem, index) => ctx.db.updateItem({
+			where: {id: sectionItem.id},
+			data: {position: itemIndex + index},
+		})),
+	);
+
+	return removedItem;
 };
 
 module.exports = {
